dynamic-expert-all

+++ expertise/src/main/java/com/vatunisia/soh/expertise/entity/Expertise.java
@@ -21,4 +21,4 @@ public class Expertise {
     private String description;
     private String minProposedPrice;
     private String maxProposedPrice;
-    private Long businessId; }
+    private Integer businessId; }
